+++ src/templates/contributors.jsx
@@ -1,6 +1,5 @@
 import { graphql } from 'gatsby';
 import { getSrc } from 'gatsby-plugin-image';
-import { axios } from 'helpers/axios';
 import React from 'react';
 
 import Achievments from 'components/pages/contributors/achievments';
@@ -11,7 +10,7 @@ import GetStarted from 'components/shared/get-started';
 import Layout from 'components/shared/layout';
 import Separator from 'components/shared/separator';
 
-const ContributorsPage = ({ serverData: { issues }, data: { ogImage }, pageContext }) => {
+const ContributorsPage = ({ data: { ogImage }, pageContext }) => {
   const SEO = {
     title: 'Novu - Contributors',
     description:
@@ -25,7 +24,7 @@ const ContributorsPage = ({ serverData: { issues }, data: { ogImage }, pageConte
       <Hero />
       <Achievments list={pageContext.contributors.list} />
       <HowItWorks />
-      <Issues issues={issues} />
+      <Issues issues={pageContext.issues} />
       <GetStarted />
       <Separator backgroundColor="black" />
     </Layout>
@@ -43,17 +42,3 @@ export const pageQuery = graphql`
     }
   }
 `;
-
-export async function getServerData() {
-  try {
-    const issues = await axios.get(`/issues`);
-
-    return {
-      props: {
-        issues: issues.data.issues,
-      },
-    };
-  } catch (err) {
-    console.log(err);
-  }
-}
